requirements.txt
Pillow==11.1.0
pygame==2.5.2
pygame==2.6.1
pynput==1.7.7
PyQt5==5.15.11
PyQt5_sip==12.16.1
pystray==0.19.5
screeninfo==0.8.1
